refactor: consolidate AI domain utilities
- Merge parse.ts (stripCodeFences, 4 lines) into config.ts — single
  import point for AI config and parsing utilities
- Merge usage-retention.ts (cleanupExpiredLogs, 37 lines) into
  usage-log.ts — co-locates all AI logging/retention logic
- Update test mocks to include stripCodeFences in ai/config mock
- Delete 2 files, reducing AI domain from 6 files to 4
