+++ apps/web/src/lib/server/domains/ai/config.ts
@@ -27,3 +27,8 @@ export function getOpenAI(): OpenAI | null {
   }
   return openai
 }
+
+/** Strip markdown code fences that some models wrap around JSON responses. */
+export function stripCodeFences(text: string): string {
+  return text.replace(/^```(?:json)?\s*\n?/i, '').replace(/\n?```\s*$/i, '')
+}
