--- apps/web/src/lib/server/domains/ai/parse.ts
@@ -1,4 +1,0 @@
-/** Strip markdown code fences that some models wrap around JSON responses. */
-export function stripCodeFences(text: string): string {
-  return text.replace(/^```(?:json)?\s*\n?/i, '').replace(/\n?```\s*$/i, '')
-}
