+++ apps/web/src/lib/server/domains/feedback/pipeline/__tests__/extraction.service.test.ts
@@ -81,6 +81,7 @@ const mockOpenAI = {
 
 vi.mock('@/lib/server/domains/ai/config', () => ({
   getOpenAI: vi.fn(() => mockOpenAI),
+  stripCodeFences: vi.fn((s: string) => s.replace(/^```[\s\S]*?\n/, '').replace(/\n```$/, '')),
 }))
 
 vi.mock('@/lib/server/domains/ai/retry', () => ({
