+++ apps/web/src/lib/server/domains/feedback/pipeline/__tests__/quality-gate.service.test.ts
@@ -14,6 +14,7 @@ const mockOpenAI = {
 
 vi.mock('@/lib/server/domains/ai/config', () => ({
   getOpenAI: vi.fn(() => mockOpenAI),
+  stripCodeFences: vi.fn((s: string) => s.replace(/^```[\s\S]*?\n/, '').replace(/\n```$/, '')),
 }))
 
 vi.mock('@/lib/server/domains/ai/retry', () => ({
