+++ apps/web/src/lib/server/domains/summary/summary.service.ts
@@ -6,9 +6,8 @@
  */
 
 import { db, posts, comments, eq, and, or, isNull, ne, desc, sql } from '@/lib/server/db'
-import { getOpenAI } from '@/lib/server/domains/ai/config'
+import { getOpenAI, stripCodeFences } from '@/lib/server/domains/ai/config'
 import { withRetry } from '@/lib/server/domains/ai/retry'
-import { stripCodeFences } from '@/lib/server/domains/ai/parse'
 import type { PostId } from '@quackback/ids'
 
 const SUMMARY_MODEL = 'google/gemini-3.1-flash-lite-preview'
